requirements.txt
Django==4.2.29
djangorestframework==3.15.2
django-filter==24.3
